add DeployerMethodClass into a separate file
